--- a/packages/rust-core/crates/engine-cli/Cargo.toml
+++ b/packages/rust-core/crates/engine-cli/Cargo.toml
@@ -19,6 +19,7 @@ once_cell = "1.19"
 
 [dev-dependencies]
 criterion = { version = "0.5", features = ["html_reports"] }
+serial_test = "3.2"
 
 [features]
 default = []