# CP-2: False-sharing対策の実装

## 概要

Phase 3のCP-2として、並列探索におけるfalse-sharing問題を解決するため、SharedHistoryとDuplicationStatsの最適化を実施しました。

## 実装内容

### 1. SharedHistoryの最適化

**変更前**:
```rust
pub struct SharedHistory {
    table: Vec<AtomicU32>,  // 2430個のエントリが連続配置
}
```

**変更後**:
```rust
use crossbeam_utils::CachePadded;

pub struct SharedHistory {
    table: Vec<CachePadded<AtomicU32>>,  // 各エントリがキャッシュライン単位でパディング
}
```

### 2. DuplicationStatsの最適化

**変更前**:
```rust
pub struct DuplicationStats {
    pub unique_nodes: AtomicU64,
    pub total_nodes: AtomicU64,  // 同じキャッシュライン上
}
```

**変更後**:
```rust
pub struct DuplicationStats {
    pub unique_nodes: CachePadded<AtomicU64>,
    pub total_nodes: CachePadded<AtomicU64>,  // 別々のキャッシュラインに配置
}
```

## 技術的詳細

### False-sharingとは

複数のスレッドが異なる変数にアクセスしているにも関わらず、それらの変数が同じキャッシュライン（通常64バイト）上にあるため、キャッシュの無効化が頻繁に発生し、パフォーマンスが低下する現象です。

### CachePaddedの仕組み

`crossbeam_utils::CachePadded<T>`は、値を64バイト境界にアライメントすることで、各値が独立したキャッシュライン上に配置されることを保証します。

```rust
// 内部実装のイメージ
#[repr(align(64))]
pub struct CachePadded<T> {
    value: T,
    _padding: [u8; 64 - size_of::<T>()],
}
```

## パフォーマンスへの影響

### 期待される効果

1. **スケーラビリティ向上**: 4スレッド以上での性能改善
2. **キャッシュミス削減**: false-sharingによるキャッシュ無効化の減少
3. **レイテンシ改善**: 共有データアクセスの競合削減

### トレードオフ

- **メモリ使用量増加**: SharedHistoryが16-32倍に増加（2430 * 4バイト → 2430 * 64-128バイト）
- 実際の増加量: 9,720 bytes (≈9.5 KiB) → 155,520-311,040 bytes (≈152-304 KiB)
- 注: CachePaddedのサイズはシステムのキャッシュライン（64または128バイト）に依存

## テストとベンチマーク

### テスト

- 既存の並列探索テストがすべて成功
- Thread Sanitizer警告なし

### ベンチマーク

`false_sharing_bench.rs`を追加し、以下の項目を測定：
- 並列SharedHistory更新のスループット
- 並列DuplicationStats更新のスループット
- History aging操作のパフォーマンス

## 今後の最適化案

メモリ使用量が問題になる場合の代替案：

```rust
// 手動パディング版（メモリ効率重視）
#[repr(align(64))]
struct PaddedAtomic {
    value: AtomicU32,
    _padding: [u8; 60], // 64 - 4 = 60
}
```

## レビュー対応（Must-fix/Should-fix）

### Must-fix対応

1. **SharedHistory::updateのCASループ実装**
   - 失敗時に1回で諦めていた問題を修正
   - 成功するまでリトライするループに変更
   - 衝突率が高い局面（王手/必至探索）でもボーナスが確実に適用される

### Should-fix対応

2. **CachePaddedサイズのアサーション追加**
   - `test_cache_padded_size`テストを追加
   - 64バイト以上のアラインメントを保証
   - 128バイトキャッシュラインのシステムにも対応

3. **DuplicationStats::get_duplication_percentage改善**
   - 式を読みやすく修正（単項演算を回避）
   - `((total - unique) as f64) * 100.0 / (total as f64)`

### Nice-to-have対応

4. **メモリ増分の正確な記載**
   - 2430 * 64-128 = 155,520-311,040 bytes (≈152-304 KiB)
   - システムのキャッシュラインサイズに依存することを明記

## 期待されるパフォーマンス向上（レビューより）

| スレッド | BEFORE NPS | AFTER NPS | Speed-up |
|---------|------------|-----------|----------|
| 1       | 1.00×      | 1.00×     | –        |
| 4       | 2.41×      | 2.71×     | +12%     |
| 8       | 4.18×      | 4.75×     | +13%     |

- false-sharing解消だけで重複率が約3ポイント減少
- 4スレッド以上でNPSが目に見えて向上

## 参考資料

- [Intel 64 and IA-32 Architectures Optimization Reference Manual](https://www.intel.com/content/www/us/en/developer/articles/manual/64-ia-32-architectures-optimization-manual.html)
- [crossbeam-utils documentation](https://docs.rs/crossbeam-utils/)